Version 8.8.3

+++ Trabajo grupal/Clases/Datosbasedt.cs
@@ -40,7 +40,7 @@ namespace Trabajo_grupal.Clases
             Conexion.cerrarcon();
             return retorno;
         }
-        public static int Eliminar(Int64 pId)
+        public static int Eliminar(int pId)
         {
             int retorno = 0;
             Conexion.opoencon();
